Fixed broken build on 32-bit systems

+++ src/gdbserver/gdb-server.c
@@ -862,7 +862,7 @@ static int flash_go(stlink_t *sl) {
     for(struct flash_block* fb = flash_root; fb; fb = fb->next) {
         DLOG("flash_do: block %08x -> %04x\n", fb->addr, fb->length);
 
-        for(stm32_addr_t page = fb->addr; page < fb->addr + fb->length; page += FLASH_PAGE) {
+        for(stm32_addr_t page = fb->addr; page < fb->addr + fb->length; page += (uint32_t)FLASH_PAGE) {
             unsigned length = fb->length - (page - fb->addr);
 
             //Update FLASH_PAGE
